--- feature-dogs/src/main/java/feature/data/remoteDataSource/RemoteDataSource.kt
@@ -1,7 +1,0 @@
-package feature.data.remoteDataSource
-
-import feature.domain.model.BreedDomain
-
-internal interface RemoteDataSource {
-    suspend fun getBreedsByPage(page: Int): List<BreedDomain>
-}
