--- feature-dogs/src/main/java/feature/domain/repository/DogsApiRepository.kt
@@ -1,7 +1,0 @@
-package feature.domain.repository
-
-import feature.presentation.model.BreedPresentation
-
-internal interface DogsApiRepository {
-    suspend fun getBreeds(page: Int): List<BreedPresentation>
-}
