+++ feature-dogs/src/main/java/featuredogs/data/api/DogsApiService.kt
@@ -1,0 +1,20 @@
+package featuredogs.data.api
+
+import featuredogs.data.response.BreedResponse
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+internal interface DogsApiService {
+
+    @GET("breeds")
+    suspend fun getBreedsByPage(
+        @Query(
+            value = "page",
+            encoded = true
+        ) page: Int,
+        @Query(
+            value = "limit",
+            encoded = true
+        ) limitOfBreeds: Int
+    ): List<BreedResponse>
+}
