+++ feature-dogs/src/main/java/featuredogs/data/remoteDataSource/RemoteDataSource.kt
@@ -1,0 +1,7 @@
+package featuredogs.data.remoteDataSource
+
+import featuredogs.domain.model.BreedDomain
+
+internal interface RemoteDataSource {
+    suspend fun getBreedsByPage(page: Int): List<BreedDomain>
+}
