+++ feature-dogs/src/main/java/featuredogs/data/remoteDataSource/RemoteDataSourceImpl.kt
@@ -1,0 +1,17 @@
+package featuredogs.data.remoteDataSource
+
+import featuredogs.data.api.DogsApiService
+import featuredogs.data.mapper.BreedsMapper
+import featuredogs.domain.model.BreedDomain
+
+internal class RemoteDataSourceImpl(
+    private val api: DogsApiService
+) : RemoteDataSource {
+    override suspend fun getBreedsByPage(page: Int): List<BreedDomain> {
+        return BreedsMapper.toBreedsDomain(api.getBreedsByPage(page, FIFTY_BREEDS))
+    }
+
+    companion object {
+        const val FIFTY_BREEDS = 50
+    }
+}
