+++ feature-dogs/src/main/java/featuredogs/domain/repository/DogsApiRepository.kt
@@ -1,0 +1,7 @@
+package featuredogs.domain.repository
+
+import featuredogs.presentation.model.BreedPresentation
+
+internal interface DogsApiRepository {
+    suspend fun getBreeds(page: Int): List<BreedPresentation>
+}
