--- feature-examples/src/main/java/feature/presentation/model/ExampleModel.kt
@@ -1,8 +1,0 @@
-package feature.presentation.model
-
-/**
- * This data class is just a class to have data as attributes.
- */
-data class ExampleModel(
-    val title: String? = null
-)
