--- feature-examples/src/main/java/feature/utils/ExampleClickListener.kt
@@ -1,10 +1,0 @@
-package feature.utils
-
-/**
- * This just a simple interface that we are gonna implement a method to have the position of the
- * element clicked in the list.
- */
-interface ExampleClickListener {
-
-    fun onClick(position: Int)
-}
