+++ feature-examples/src/main/java/featureexamples/presentation/model/ExampleModel.kt
@@ -1,0 +1,8 @@
+package featureexamples.presentation.model
+
+/**
+ * This data class is just a class to have data as attributes.
+ */
+data class ExampleModel(
+    val title: String? = null
+)
