+++ feature-examples/src/main/java/featureexamples/utils/ExampleClickListener.kt
@@ -1,0 +1,10 @@
+package featureexamples.utils
+
+/**
+ * This just a simple interface that we are gonna implement a method to have the position of the
+ * element clicked in the list.
+ */
+interface ExampleClickListener {
+
+    fun onClick(position: Int)
+}
